0323 - 파일 처리증

+++ javajava/src/main/java/com/example/javajava/board/service/BoardService.java
@@ -16,7 +16,7 @@ public interface BoardService {
 
     void boardUpdate(BoardDto board);
 
-    void boardDelete(int boardIdx);
+    void boardDelete(Long boardIdx);
 
-    FileDto selectFileInfo(int idx, int boardIdx);
+    FileDto selectFileInfo(Long idx, Long boardIdx);
 }
